+++ react-app/src/components/Navigation/navbar.css
@@ -3,7 +3,7 @@
     grid-template-columns: repeat(10, 10vw);
     width: 100%;
     height: 100%;
-    grid-row: 1 / 4;
+    grid-row: 1 / 3;
     grid-column: 1 / end;
     background-color: #1f539288;
