--- react-app/src/components/splashpage/Splashpage.js
@@ -1,13 +1,0 @@
-import {Link} from 'react-router-dom';
-import './splashpage.css'
-function Splashpage() {
-    return(
-        <div className="wrapper">
-            <div className="about-us"> about us</div>
-            <div className="feature-chore"> feature chore</div>
-            <div className="about-me-links">test</div>
-        </div>
-    )
-}
-
-export default Splashpage;
