--- react-app/src/components/splashpage/splashpage.css
@@ -1,16 +1,0 @@
-
-.wrapper {
-    display: flex;
-    background-color: lightblue;
-    justify-content: space-around;
-    height: 100vh;
-}
-.about-us {
-    display: flex;
-    background-color: blue;
-}
-
-.feature-chore {
-    display: flex;
-    background-color: red;
-}
