+++ react-app/src/index.css
@@ -2,15 +2,6 @@
 /* body {
   background-color: #482525;
 } */
-html, body {
-	margin: 0;
-	padding: 0;
-	border: 0;
-	font-size: 100%;
-	font: inherit;
-	vertical-align: baseline;
-}
-
 /* * this is css reset */
 html, body {
 	margin: 0;
@@ -22,6 +13,7 @@ html, body {
 }
 
 #root {
+  overflow: hidden;
   display: grid;
   height: 100vh;
   width: 100%;
